requirements.txt
PyQt5>=5.15
Pillow>=8.0
# Optional (for future hash-based matching)
imagehash>=4.2